CP-45566: [1/4] Change to use new guidance format in updateinfo.xml
The guidance in new format in updateinfo.xml file deprecates the
guidance in old format. For backwards compatibility support, both
old and new guidance data will be in updateinfo.xml.

With this commit, XAPI uses the guidance in new format only as the data
in new format has been presented in updateinfo.xml.

Signed-off-by: Ming Lu <ming.lu@cloud.com>