CP-45566: [2/4] Update evaluating guidances from new data structures
This commit updates the guidance evaluation logic and exposes the result
of evaluation to the response of query on HTTP /updates endpoint.

The response will be consumed by clients like XenCenter. The format of
response is changed in this commit. Regarding this change, the backwards
compatibility is not supported on this interface.

Additionally, the evaluation is simplified by not considering the failed
livapatches any more. Otherwise, the client like XenCenter has to
introduce complicated UI design to populate the guidance change due to
the livepatch failures. This may cause un-necessary carrying out of
livepatch guidances. But the overall impact of carrying out these
guidances is considered as ignorable.

Signed-off-by: Ming Lu <ming.lu@cloud.com>