CA-383867: Add local disk cache library for xapi guard
This enables xapi-guard to decouple persistence of TPM contents from the xapi
service being online. That is, when xapi is down. The contents of the TPMs will
be written to disk, and when xapi is back online the contents will be uploaded.

This is needed to protect VMs while xapi is being restarted, usually as part of
an update.

Some properties of the cache:
- The cache is tried to be bypassed whenever possible, and is only used as
  fallback after a write fails.
- The cache is handled by a thread that writes to cache and one that reads from
  it. They communicate through a bounded queue.
- Whenever a TPM content is written to disk, previous versions of it are
  deleted. This helps the reading thread to catch up.
- When the queue has been filled the writer stops adding elements to the queue,
  and the reader will try to flush the queue, and after it will try to flush
  the cache. After this happens both threads will transition to cache bypass
  operation.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>