CA-384457: Add special case for End_of_file to sparse_dd_wrapper
Previously sparse_dd wrapper would fail to read the progress from
sparse_dd when it exits abnormally, e.g. due to an write error to the
underlying vdi, in which case the error message gets changed to
End_of_file because the wrapper cannot read the progress. This patch
adds a special case when we encounter End_of_file and preserves the
error message from sparse_dd.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>