CP-45567: Set pending RestartVM for all VMs in the pool
RestartVM is a task specified in update guidance. It has to be done on
an updated host. Unlike other VM tasks in update guidance, a VM may
escape from the tracking on individual host. E.g. a host evacuation will
migrate a VM to aother host before host.apply_updates call setting the
pending guidance on it. It's also not good to expose an API to set the
pending guidance lists by clients.

The solution in this commit is to set it for all running/paused/suspended
VMs in the pool when the host.apply_updates is called for the coordinator
host which is always the first one to be updated in a pool update.

Signed-off-by: Ming Lu <ming.lu@cloud.com>