CA-386582: Always create exporting thread for observers
Previously the exporting thread for observer is only created when there
is an observer stored in the database associated with a component. This
is not the original intent of how observers should be used.

We want to create the exporting thread regardless of whether we have any
observers so that when observers are created later on, their spans can be
exported, since creating an observer will not check for the existence of
an exporting thread.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>