CP-45570: Clear host update guidance
There were 2 problems about host update guidance:

1. RestartToolstack guidance in pool member was cleared when there was
a xapi restart in pool coordinator.

This is because "consider_enabling_host_request" in pool member would be
called when there is a xapi restart in pool coordinator, as the db
connection to the coordinator will be re-established then.

2. CA-387695: RebootHost guidance in pool member got cleared when there
was a xapi restart in pool coordinator.

When xapi is started with a host (re)boot, its command line looks like:
/opt/xensource/bin/xapi -nowatchdog
                        -writereadyfile /var/run/xapi_startup.cookie
                        -writeinitcomplete /var/run/xapi_init_complete.cookie
                        -onsystemboot
then "!Xapi_globs.on_system_boot" will always be true.

Same as above, "consider_enabling_host_request" in pool member would be
called when there is a xapi restart in pool coordinator. And when
"!Xapi_globs.on_system_boot" is true for xapi in the pool member, the
RebootHost guidance in pool member would be cleared when there was a
xapi restart in pool coordinator.

These 2 problems are resolved by clearing host update guidance in
function "Dbsync_slave.update_env", which is called only during xapi
startup.

Signed-off-by: Gang Ji <gang.ji@citrix.com>