CP-47046 optimise rule emitter for xenopsd PVS Proxy setup
Emit port-oriented rules for IPv4 and IPv6 only when needed.

The complication is that a single PVS server can have multiple addresses
each of which can be IPv4 or IPv6. So we need both kind of rules if
addresses as mixed but can emit only one kind if addresses are of one
kind.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>