CP-46909: Update error returned from HTTP endpoint /updates
The error was designed as HTTP error code only. This is not friendly
to clients like XenCenter, i.e. XenCenter can't get more
information from the error. The improvement is to return a JSON format
data like:
{
    "error": {
        "code": 1,
        "data": ["<data1>", "<data2>", ...],
        "message": "<XAPI error code>"
    }
}.
The data list are the parameters of the error defined in datamodel_error,
and the HTTP response code will be 200 always.

Signed-off-by: Gang Ji <gang.ji@citrix.com>