Fixup: filter out EvacuateHost before safety check
A safety check in host.apply_updates will raise error if there is a
mandatory EvacuateHost guidance but some VMs are running on the host.
Meanwhile, EvacuateHost should not be added into pending list as it
should be be carried out before host.apply_updates. Hence it needs
to be filtered out.

This commit fixes a bug that filering out the EvacuateHost before the
safety check. It should be filtered out after the safety check.

Signed-off-by: Ming Lu <ming.lu@cloud.com>