build: add command to check that the C# SDK compiles
We need a workaround to compile .resx on Linux (it currently relies on a windows-only build tool): just pretend its null.

We also need to add a dependency on the Newtonsoft Nuget (this should be added to the SDK itself once it fully switches to the unpatched version).

Not yet integrated into the CI, that will be done by a followup commit.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>