CA-387201: Pool.last_sync_date not reset if the user changes the update channel
When a user adds more update repositories as enabled ones, the
pool.last_update_sync will be invalid as at the moment not all the
enabled repositories have been synced at the time which is recorded
in pool.last_update_sync.

To avoid confusion, this commit resets the field to epoch so that the
user can be notified to do a pool.sync_updates soon.

Disabling an update repository will not make the pool.last_update_sync
be invalid unless no enabled repositories after the disabling.

Signed-off-by: Ming Lu <ming.lu@cloud.com>