fix(test): use SO_REUSEADDR to rebind port when run in a loop
When running 'dune runtest --no-buffer --force' in a loop sometimes we fail to bind with an EADDRINUSE error.
This is expected behaviour with TCP: when a process closes its listening socket it enters TIME_WAIT state and cannot be reused for a while.
This can be seen by running `ss -na | grep 9411` which shows TIME_WAIT.

However this is test code, and we have no remote clients that needs to receive FIN/ACK: any clients we had have already finished.
SO_REUSEADDR will only fail to bind if there is an application actively listening on it, but otherwise allow binding the port,
bypassing TIME_WAIT.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>