py3: make sure we are not using unicode type in python3
This is a redo of 48c8c3ec425c89afcd839b6a9c5b2bf7725de567, not
touching `long` and `str`.

`unicode` is only used in there for testing whether we have a string,
so aliasing it to `str` is valid.  OTOH we likely don't want to accept
`bytes` where we accept `str` in python2, and `str` gets aliased to
`bytes` in other areas of the code, so this might reveal issues in
other places.

Signed-off-by: Yann Dirson <yann.dirson@vates.fr>