fix(ci): remove 1024 fd limit for now
Setting this is useful for speeding up tests (by default my container has 1048576 as ulimit which makes the tests very slow),
but there is something non-deterministic in the environment GitHub Actions run in:
* the tests pass on a PR, but fail when run as a scheduled CI job on master

We probably don't estimate correctly how many file descriptors the parent process uses in the test
(GH actions seem to use more than locally).

Until we fully understand this revert the 'ulimit -n' part of this change.

Fixes: 954f3d337 ("build: set a timeout for the tests")
Signed-off-by: Edwin Török <edwin.torok@cloud.com>