[maintenance]: enable generate_opam_files in dune-project
Use the simplest solution: copy *.opam to *.opam.template so that we get exactly the same opam files as before,
and only declare the names in dune-project.

This will allow us to merge subprojects that already use dune-project to generate opam files (such as xapi-stdext),
and we can gradually move opam files to be generated by dune.
However first we should drop the opam files that we do not need,
which will be partially done in https://github.com/xapi-project/xen-api/pull/4939

Signed-off-by: Edwin Török <edwin.torok@cloud.com>