CP-46631: Forcefully finish most frequent spans
If the span table gets full, after we log the most frequent spans, we
forcefully finish the oldest half of this list. We round up to a least
one span. These are the spans that are most likely to be leaking.
Spans that are close because of this have a new attribute set:

`xs.tracing.finished.reason=gc.full`.

Adds a unit test in `test_observer.ml` to check the spans are finished
properly.

There are instances in which spans are created and never finished in
xapi. Plus, when the span table is filled, it stops accepting spans.
This commit targets to improve the debuggability of leaked spans by
forcefully closing some of the spans and marking them with a new tag.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>