CP-46324: Send alert when a host leaves/joins the cluster
We add three new alert messages in addition to the existing
`cluster_host_fencing`: `cluster_host_leaving`, `cluster_host_joining`
and `cluster_quorum_approaching_lost`.

The leave and join message are added whenever a host leaves the cluster,
whether as a result of the user operation, or due to unexpected errors
(in which case we will likely get an additional fencing alert).
The approaching quorum message is sent when the cluster cannot tolerate
any more loss of hosts.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>