CP-46586 create backup VDI with determinstic UUID
When creating a backup VDI on an SR we want to derive the VDI's UUID
from the SR. This way we can't be later tricked into accepting a
different VDI as a backup.

Implement a hash that dereives the VDI UUID and pass this UUID to SM to
be used.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>