CP-46179 create backup VDI with determinstic UUID
Backport of bbc5a1f7988d47408cb63e56e2cbdaf74f7529a7

When creating a backup VDI on an SR we want to derive the VDI's UUID
from the SR such that we know how to find it later on the SR.

Implement a hash that derives the VDI UUID and pass this UUID to SM to
be used.

Currently the creation of this VDI in Xapi is detected based on its
name-label: "Pool Metadata Backup". The SM stack usually creates a new
random UUID but will use a given UUID passed as "vdi_uuid" for its
"vdi_create" command. The difficulty in the implementation is that
"vdi_uuid" is used in other commands as well and we have to make sure to
set it only in the intended context. It is not straight forward to pass
the UUID from Xapi_vdi.create down to Sm.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>