Remove the build failure & update schema_minor_vsn
So that we can have a build from feature branch to test with before
merging the feature branch into master.

But then we will have to remember to update the "db schema version" in
"upgrade_update_guidance" in "xapi_db_upgrade.ml" if "schema_minor_vsn"
is increased again when the feature branch got merged into master.

Signed-off-by: Gang Ji <gang.ji@citrix.com>