CA-383867: Integrate the cache into xapi-guard's main loop
Now the process creates a thread to read from disk and push vtpm events to xapi
at its own pace, and integrates the disk-writing part into the callback of the
deprivileged sockets.

Special consideration was taken for the resume, when the deprivileged sockets
and the write-to-cache function need to be integrated in a different way from
the codepath that creates the sockets from the message-switch server.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>