CP-46179 update metadata backup/restore to use deterministic UUID
Backport of 65f152de687229946eaea6ddcaa5e3d0a11b2b01

We introduce deterministic UUID for backup VDIs to avoid searching for
VDIs. Add this capability to the scripts that create and restore
backups. In case the VDI with the expected UUID does not exist, let the
user confirm probing for it.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>