Add the default_attributes to Dom0ObserverConfig Observers
The default_attributes are not added to the DB and so these are missing
in the observer.conf. Dom0ObserverConfig currently lacks a single source
of truth outside of the DB (and default_attributes is not given to the
DB) and so attributes outside of the DB are lost whenever a different
setter function is used. This should be reworked in future so that
observer.conf is read from/updated instead of regenerated each time.

Signed-off-by: Steven Woods <steven.woods@citrix.com>