redo_log: report redo log as broken if we cannot find the block device
XAPI was just retrying endlessly in a loop saying "Could not find block device",
but didn't raise any alerts that XenRT could detect.

Report the redo log broken the first time we fail due to the lack of a bloc k device (which would indicate something going wrong in SM).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>