CP-46631: Better abstraction for span tables.
Hides the implementation of span tables behind a `SpanTbl` data structure
that is meant to share common functionality between the `ongoing_spans`
table and `finished_spans` table.

This is meant as a follow up to a previous PR to prevent issues with
double locking when transitioning spans from `ongoing_spans` to
`finished_spans`.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>