CP-46151: Productise the observer.py.
Instead of being a noop, it will now automatically instrument the python
code given to it and create spans accordingly. The trace created will
link to the xapi code that called it. It will export to all endpoints
that it is given, which it reads from the observer.conf files created by
xapi.

Signed-off-by: Steven Woods <steven.woods@citrix.com>
Co-authored-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>
Co-authored-by: Marcus Granado <590521+mg12@users.noreply.github.com>