CA-383867: Add startup procedure to xapi-guard
Because timestamps depend on a monotonic timestamp that depends on boot, files
need to be renamed to ensure future writes have higher timestamps to be
considered newer and be uploaded to xapi.

On top of this, allows to report about remnant temporary files, delete invalid
files and remove empty directories.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>