CA-383867: xapi-guard's cache contents are used in read requests
For domains requesting the TPM's contents, the xapi-guards returns the contents
in the cache, if they are available from in-flight requests. It falls back to
xapi if that couldn't be possible.

The cache doesn't try to provide any availability for reads, like it does for
writes. This means that if swtpm issues a read request while xapi is offline,
the request will fail, as it happened before this change.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>