Set the unit test codcov target to 80%
There were no python3 unit tests now.
Set the unit test coverage to 80% for the py3 update.
In the future, if there are cases where the scripts are not UTable
or we can cover them by manual test or XenRT test,
we can handle it by excluding them from the check.

Signed-off-by: Stephen Cheng <stephen.cheng@cloud.com>