pre-commit: Add pytest,pytype-analysis: Configure for quick local CI tests
This adds the configuration for the command "pre-commit run pytest -av"
to run the pytest with the same coverage checks that run in GitHub CI.

It is no longer needed to push to GitHub get get an authoritative answer
on whether CI will pass.

Installation:
$ pip install pre-commit

For relatively quick tests (pytest, pylint etc), run:
$ pre-commit run pytest -av

For all checks, including pytype, run
$ Run pre-commit run pytest -av --hook-stage push

It can optionally also also be installed as commit and push hook.
See the header of .pre-commit-config.yaml for more information.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>