CA-390109: Use `$PROFILE` path to store and read known cert list
Before these changes, the `SaveCertificates` method relied on machines having a `SpecialFolder.MyDocuments` folder. This is true in Windows and GUI versions of some Linux distros, but it's an assumption that caused the save method to fail if the folder didn't exist.

With this commit, we're storing and reading from the path where `$PROFILE` is stored, which is platform agnostic from the point of view of the SDK.

Signed-off-by: Danilo Del Busso <danilo.delbusso@cloud.com>