API docs in Hugo
This is mostly copied from the jekyll version at xapi-project.github.io and
amended for Hugo. The main difference is that all XenAPI pages have now been
integrated with the main menu in the left side bar, including the menu of the
XenAPI classes.

Some automation is left to do in order to take the `doc/data` files from a
build of `ocaml/idl` and update the class pages using the script
`doc/make-class-pages.py`. Also the overview of xapi releases is still missing,
as well as some of the API guides.

Finally, there is some overlap with markdown files in `ocaml/doc`, which needs
to be sorted out.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>