CA-390277: Reduce record usage on CLI cross-pool migrations
Using records in cross-pool migration code is dangerous, as the code interacts
with potentially newer hosts. This means that fields in the record might be
different from what's expected. In particular adding an enum field can break
the deserialization, and removing a field as well.

With this change, only SR records are used. This is done to minimize the number
of calls done.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>