CP-44752: SDK(C#): Conditionally propagate W3C traceparent
This is not available on .Net 4.5.
Could use .Net Standard 2.0, .Net Framework 4.6.2 and .NET 8.0 as target,
which matches the availability of the
System.Diagnostics.DiagnosticSource builtin (and Nuget package for older
versions):
```
NET462_OR_GREATER || NETSTANDARD2_0_OR_GREATER || NET8_0_OR_GREATER
```

However there are some bugs in the CI where the xenserver-samples XenSdkSample ends up
in a zip, but without its dependency (the DiagnosticSource.dll).

So for now just enable this on .Net8, where hopefully this package is part of the runtime already
(although of course you'll still need to install the runtime, which for some reason you didn't have to do
with .Net6, perhaps .Net6 would come preinstalled on Debian12?).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>