CA-391880: Update related field 'groups' of VM when destroying VM group.
In VM anti-affinity datamodel PR (https://github.com/xapi-project/xen-api/pull/5573),
there is many-many relation between VMs and VM groups.
 When changing the VM.groups manually, the VM_group.VMs will be updated automatically.
But when removing a VM group, the VM.groups will not be updated automatically.
This commit aims to update VM.groups for all VMs in a VM group when it is being removed.

Signed-off-by: Bengang Yuan <bengang.yuan@cloud.com>