python3/unittest: Replace import_file with import_helper
Bernhard Kaindl: Resolved a conflict with one of my other
upcoming changes which removes the need to a pytype ignore by
using a cast() instead, which fixes pyright/pylance/vscode.

Also keep import_helper in python3/tests as the the name
"unittest" sould likely avoided as it clashes with the unittest
module.

The tests in python3/tests are written for pytest instead and
not all of them will classify as unit tests, so using the name
unittest for them would also be a misnomer.

Using the name tests is shorter and more generic, and we can
also use the separation beween tests and unittest to differentiate
between modern pytest tests and legact unittest-based tests which
should possibly be better migrated to pytest at some point for the
benefits that pytest gives: For example with pytest, you can use
use just assert and you do not need to use self.assert...(),
because pytest implements the proper assert matching diagnostics.

In the long run, the classic unittest tests should no longer
be used.

Co-authored-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>
Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>