ci(nopin): pinning is very slow and not necessary
We do not publish or use the opam packages from this repo,
and therefore pinning is unnecessary.
'opam install --deps-only --with-test .' works without pinning on the version of 'opam' we have here.

Pinning ~71 packages is very slow since it makes a full copy of the source+.git folder every time.
Skip pinning, this also saves a few gigabytes of disk space.

This also makes the cleanup-xapi-environment action a no-op, so remove it.
(in fact it'd fail because the output of opam list with master pins is empty).

For testing the correctness of opam packages:
* we still have the xs-opam CI
* we can introduce opam-dune-lint that statically checks opam and dune for missing dependencies (needs new release first)

This reduces CI times:
* Run OCaml tests: from 15m24s to 10m16s
  * Setup XenAPI environment: from 7m1s to 5m55s

Signed-off-by: Edwin Török <edwin.torok@cloud.com>