ci: trim dune cache
We should eventually separate the opam dependency install into its own job
that uploads an artifact/cache that other jobs can reuse (and that should have a bigger dune cache),
however for now just trim the size of the cache.

The default is 5GiB, which results in ~1.4GiB compressed, which would hit the 10GiB cache size limit in GitHub way too often.
We may need to tweak this value in the future.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>