CP-49140: [prep]: database: drop unused of_sexp conversions
sexp_of is used when saving the schema, but of_sexp is unused.

This enables a future commit to store the deserialized value as a pair value,
and deserialization (unmarshaling) requires knowing the schema,
whereas serialization (marshaling) does not.
Thus we couldn't implement a generic `t_of_sexp` function, but with this change
we won't have to.

No functional change.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>