CP-49633: Pass traceparent for non-local `XenAPI.py` sessions.
`transport.make_connection` is now patched to pass a `traceparent`
header when creating a non-local session.

e.g. `XenAPI.Session(uri, transport)`

This is meant as a follow up to e5bb639b that applied only to
UDSTransport (local_login). It is  more generic and it applies
to anything derived from Transport, including UDSTransport,
SafeTransport and Transport.

The new behaviour is only present if otel=true, i.e. OTEL_SDK_DISABLED
is set to `false` and required open-telemetry libs are present.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>