CP-49707: Go SDK: Add JSON tag for structs
This aims to allow a Go SDK client to unmarshal JSON to Go structs with
the function in Go json pacakge instead of using Go SDK.

The unmarshalling in Go SDK contains two steps: JSON -> map -> struct.
This is because some special cases need to be handled explicitly.
But for sake of safety, this unmarshalling is not exposed to clients.

Signed-off-by: Ming Lu <ming.lu@cloud.com>