Go SDK: Misc fixes for on-going component tests (#5661)
* CP-49707: Go SDK: Add JSON tag for structs

This aims to allow a Go SDK client to unmarshal JSON to Go structs with
the function in Go json pacakge instead of using Go SDK.

The unmarshalling in Go SDK contains two steps: JSON -> map -> struct.
This is because some special cases need to be handled explicitly.
But for sake of safety, this unmarshalling is not exposed to clients.

Signed-off-by: Ming Lu <ming.lu@cloud.com>

* CP-49349: Go SDK: Append UTC timezone to unix timestamp

The date time in Go requires a date time with timezone info. But XAPI
may return a UNIX timestamp to clients.

This commit blindly adds a UTC timezone for UNIX timestamps returned
from XAPI server.

Signed-off-by: Ming Lu <ming.lu@cloud.com>

* CP-49349: Go SDK: Allow spaces in special value of float type

Some special values of float standing in JSON are handled specifically.
This commit allows the special values contain spaces.

Signed-off-by: Ming Lu <ming.lu@cloud.com>

* Go SDK: Disable golangci tagliatelle

This check complains the struct json tag naming style. It expects the
Caml style, but Go SDK has to use Snake.

Signed-off-by: Ming Lu <ming.lu@cloud.com>

---------

Signed-off-by: Ming Lu <ming.lu@cloud.com>