CA-393507: Default cluster_stack value
Add a `@default` attribute to the `cluster_stack` field in the cluster
config. This field was introduced in #c1bd0e31a but causes RPU to fail
since it is not optional, and while xapi-clusterd was unmarshalling an
old db (which contains cluster config), it cannot find this field, and
throws an exception.

Adding `@default` can make the rpc library to fill this field in when it
is missing, solving the above problem.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>