CP-49249: Implement SMAPIv3 CBT Forwarding
Implements the changes required to forward CBT-related functionality to the
SMAPIv3 storage plugins, which now support CBT for XFS and GFS2 storage
repositories.

The length of the changed blocks requested is hardcoded as a negative number,
which is interpreted by the related plugins as requesting the changes for the
entire length of the VDIs being compared. SMAPIv1 does not support this
functionality but we retain it here with this behaviour in case we want to make
the functionality explicit in future.

Signed-off-by: Colin James <colin.barr@cloud.com>