CP-49129: Replace `ocamlyacc` with `menhir`
From https://ocaml.org/manual/5.1/lexyacc.html on ocamlyacc concurrency
sefety:

"Parsers generated by ocamlyacc are not thread-safe. Those parsers rely
on an internal work state which is shared by all ocamlyacc generated parsers.
The menhir parser generator is a better option if you want thread-safe parsers."

We currently hold a global lock while calling the sexpr parser, so even if we
try to parse a small sexpression, it'll be blocked behind parsing the large one.

Switch to Menhir to make this thread-safe.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>