CP-49858: Remove template variables 'first' and 'is_session_id'
Prior to this change, the function parameter list is generated by
Mustache template. As a result, a template boolean variable 'first' has
to be used to determine if a leading comma is presented or not. E.g.
when the 'first' is true, a parameter would be rendered as "a string',
and then the next parameter with the 'first' being false would be
rendered as ", b string". Putting them together would result in
"a string, b string".

This could work but is difficult to be understood.

In this commit, it is changed to construct the whole function parameter
list as a string in OCaml code and bind it to the template variable
'func_params'.

Similar changes apply to the template variable 'is_session_id'. It is
removed in this commit and the expected values are generated by OCaml
code.

Signed-off-by: Ming Lu <ming.lu@cloud.com>