CA-394169: Allow task to have permissions on itself
Backport of f25cee63e

Do not check permissions of a task if it is acting on itself.
In the case the session is lost, the task needs to be able to update
itself.

If the session were to be lost, we would not able to retrieve back the
correct `auth_user_id`. If the task ids match implies that we know we
were already authorized.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>

Co-authored-by: Christian Lindig <christian.lindig@cloud.com>