CA-392887: set_tls_config immediately after enabling clustering
Previously xapi calls `set_tls_config` regardless of whether a host
has joined or enabled, which will restart the remote server of
xapi-clusterd. In the meantime, another xapi-clusterd might also be
joining, which causes `distribute_state` to be called while the
remote server is restarting.

Now remove the `set_tls_config`, this is because `join_internal` already
creates a tls_config and passes it to xapi-clusterd, but xapi-clusterd
does not store that tls_config in its db, it just starts the http server
with that tls config. Modifying xapi-clusterd to store that config will
be done in a separate PR. Moreover, `cluster_host.enable` also calls `set_tls_config`,
which means there is no need to call `set_tls_config` if the cluster
host is joined but not enabled.

Also move the observer and watcher creation into the not joined case,
since cluster_host.enable already calls them and there is no need to
call them if the host is not enabled.

This does not, however, solve the whole problem. For that, we need to
make sure that `distribute_state` and `set_tls_config` cannot happen at
the same time. More generally, any remote calls cannot happen while
`tls_config` is running. Hence we need them to hold the same lock. This
will be done in xapi-clusterd.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>